Update README.md
This is a C++ code for a simulated lottery. First the user inputs 5 numbers (which can be thought of as their assigned tickets for the big draw). Based on the tickets, randomized 5 digit codes are generated per user. If the 5 number combination pulled out from the lottery matches exactly with the user's code, then that user is the lucky lottery winner!